+++ package.json
@@ -39,6 +39,7 @@
     "joi": "^17.9.2",
     "jsonwebtoken": "^9.0.0",
     "reflect-metadata": "^0.1.13",
-    "type-graphql": "^2.0.0-beta.2"
+    "type-graphql": "^2.0.0-beta.2",
+    "typedi": "^0.10.0"
   }
 }
