+++ src/dtos/inputs/create-user-input.ts
@@ -19,7 +19,4 @@ export class CreateUserInput {
 
   @Field()
   confirmPassword: string
-
-  @Field()
-  adressId: number
 }
