+++ src/resolvers/signup-resolver.ts
@@ -1,12 +1,14 @@
-import { Arg, Mutation, Query, Resolver } from "type-graphql";
+import { Arg, Ctx, Mutation, Query, Resolver } from "type-graphql";
 
-import { context } from "../context";
+import { Context, context } from "../context";
 import { CreateUserInput } from "../dtos/inputs/create-user-input";
 import { User } from "../dtos/models/create-user-model";
 import { Adress } from "../dtos/models/create-adress-model";
 import { CreateAdressInput } from "../dtos/inputs/create-adress-input";
 
-import bcrypt from 'bcrypt'
+import { SignupService } from "../services/signup-service";
+
+const signupService = new SignupService()
 
 @Resolver()
 export class SignupResolver {
@@ -17,41 +19,15 @@ export class SignupResolver {
   }
 
   @Mutation(() => User)
-  async createUser(@Arg('data') data: CreateUserInput) {
+  async createUser(@Arg('data') data: CreateUserInput, @Ctx() ctx: Context): Promise<User> {
 
-    const adress = await context.prisma.adresses.findFirst({
-      orderBy: {
-        id: 'desc'
-      },
-      select: {
-        id: true
-      },
-      take: 1
-    })
-
-    if (adress) {
-      data.adressId =  adress.id
-    }
-
-    if (data.password == data.confirmPassword) {
-      const SALT = 10
-
-      const passwordEncrypted = bcrypt.hashSync(data.password, SALT)
-      const confirmPasswordEncrypted = bcrypt.hashSync(data.confirmPassword, SALT)
-
-      data.password = passwordEncrypted
-      data.confirmPassword = confirmPasswordEncrypted
-    }
-    
-    const newUser = await context.prisma.users.create({ data: data })
-
-    return newUser
+    return signupService.createUser(data, ctx)
   }
 
   @Mutation(() => Adress)
   async createAdress(@Arg('data') data: CreateAdressInput) {
 
-    const newAdress = await context.prisma.adresses.create({ data: data})
+    const newAdress = await context.prisma.adresses.create({ data: data })
 
     return newAdress
   }
